requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.3
pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.0
matplotlib>=3.8.0
python-dotenv>=1.0.1
tqdm>=4.66.0
segformer_pytorch==0.0.6
xformers>=0.0.27
scipy>=1.11.0
